- fixed test - REST API compliant with standards

+++ src/main/java/fr/polytech/al/tfc/account/model/Caps.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class Caps {
-    private float money;
-    private float amountSlidingWindow;
+    private Integer money;
+    private Integer amountSlidingWindow;
 }
